backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
plaid-python>=11.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
python-multipart>=0.0.5 